Add test to check if large files work with create().
Note that this depends on a system() call that execute fallocate to create the file, which is only on linux.  So if we are on MacOS, the test is skipped.  It also transfers a 4Gb file to the MN, so should be run somewhere with adequate bandwidth.  We probably should refine the skip() condition to be very specific to an environment we know will work, such as on jenkins.