Fixed dead code bug introduced in sha 4e1cda5bc1df9c5b0205bfdd720d87cd38278c9c
In the case where a node was not passed into the auth_get and auth_head functions, the isAuthValid() call would be skipped, and we fall through to the else. But, that block now also evaluates `node`, which produces an error. The joint condition check was a bug, so moved it to a separate check, more clearly showing that node is a required parameter (as is url).  I still have problems with passing in node here at all - the auth_* methods were not supposed to know anything about nodes. But that is a discussion for another time.