Re-enable travis package cache
This wasn't the fix after all for what I was trying to do. I disabled the cache temporarily because it looked like the travis failure was due to using an outdated version of datapack (not the newest on CRAN) but it turns out the changes we need to get rdataone's tests passing aren't on CRAN yet. So the build will fail until datapack 1.4.0 lands on CRAN.